ui-commit: add support of test results
Git SCM allows to store information in so called notes [1] and
cgit allows to show notes in in default ref ('refs/notes/commits')
attached to commits. It is convenient to attach various information
related to commit.

In some projects tests runs quickly and can be run sycnhronously before
commit, so successful results is one of mergeing criteria. But in big
projects tests there are long tests that cannot be run on pre-commit
stage and runs post-commit. In such cases it is convenient to attach test
results to easily know what tests were passed on that commit and what
tests are not. Someone may keep testing results in Git notes using
git-test [2].

How-to use:

$ git notes add -F report.tap 2a3a0d1a62de2ae5ab4511c15d82a6a0f2c2a930
$ git notes add -m 'Tested-by: Sergey Bronnikov <estetus@gmail.com>' \
	    2a3a0d1a62de2ae5ab4511c15d82a6a0f2c2a930

Open commit in cgit and download test report for that commit.

1. https://git-scm.com/docs/git-notes
2. https://github.com/ligurio/git-test
3. https://github.com/ligurio/testres/wiki/Using-GIT-as-a-storage

Signed-off-by: Sergey Bronnikov <estetus@gmail.com>